requirements.txt
Django==1.8.3
oauth==1.0.1
oauth2==1.5.211
oauth2client==1.4.12
oauthlib==2.0.0
python-social-auth==0.2.12
requests==2.11.1
requests-oauth2==0.2.0
requests-oauthlib==0.5.0
django-semanticui-forms==1.6.5